+++ app/src/Model/Menus.php
@@ -1,0 +1,14 @@
+<?php
+namespace Dappur\Model;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Menus extends Model
+{
+    protected $table = 'menus';
+    protected $primaryKey = 'id';
+    protected $fillable = [
+        'name',
+        'json'
+    ];
+}